demo ready

+++ src/main/java/com/google/sps/servlets/Delete.java
@@ -89,7 +89,7 @@ public class Delete extends HttpServlet {
       {
         try
         {
-          //if(eventsToDelete[i].deleteCalendarEvent()) 
+          if(eventsToDelete[i].deleteCalendarEvent())
             datastore.delete(keys[i]);
         }catch(Exception e)
         {
